+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # Changelog
 
-
 We are currently working on porting this changelog to the specifications in
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+
+## [Version 0.2.9] - Unreleased
+
+### Added 
+* `vimtk.Mode` with basic functionality
+* `vimtk.TextInsertor.overwrite` 
 
 ## [Version 0.2.8] - Unreleased